requirements.txt
python-binance==1.0.17
pandas==2.0.3
numpy==1.24.3
pandas-ta==0.3.14b0
scikit-learn==1.3.0
joblib==1.3.2
openpyxl==3.1.2
requests==2.31.0
